backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pymupdf
groq
sentence-transformers
faiss-cpu
tiktoken
numpy
python-dotenv
huggingface-hub
python-docx
